pass SpaceServer unit test
