+++ tests/CMakeLists.txt
@@ -25,7 +25,7 @@ IF (CXXTEST_FOUND)
 		IF (HAVE_SERVER)
 			ADD_SUBDIRECTORY (server)
 		ENDIF (HAVE_SERVER)
-
+		ADD_SUBDIRECTORY (spacetime)
  		ADD_SUBDIRECTORY (spatial)
 
 		ADD_SUBDIRECTORY (nlp)
